requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai>=1.3.0
qdrant-client>=1.7.0
langgraph>=0.0.40
python-dotenv>=1.0.0
pydantic>=2.5.0